+++ package.json
@@ -15,6 +15,7 @@
     "concurrently": "^5.3.0",
     "express": "^4.17.1",
     "multer": "^1.4.2",
-    "nodemon": "^2.0.6"
+    "nodemon": "^2.0.6",
+    "throng": "^5.0.0"
   }
 }
